fixed naming and dependencies
