+++ src/DependencyInjection/Configuration.php
@@ -84,11 +84,11 @@ class Configuration implements ConfigurationInterface
                 ->arrayNode('executor')
                     ->addDefaultsIfNotSet()
                     ->children()
-                        ->enumNode('type')->values(['inline', 'process'])->defaultValue('inline')->end()
-                        ->arrayNode('process')
+                        ->enumNode('type')->values(['inside', 'separate'])->defaultValue('inside')->end()
+                        ->arrayNode('separate')
                             ->addDefaultsIfNotSet()
                             ->children()
-                                ->scalarNode('console')->defaultValue('%kernel.root_dir%/../bin/console')->end()
+                                ->scalarNode('console_path')->defaultValue('%kernel.root_dir%/../bin/console')->end()
                             ->end()
                         ->end()
                     ->end()
