+++ src/Resources/config/executor/separate.xml
@@ -3,8 +3,9 @@
            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
     <services>
-        <service id="task.executor.inline" class="Task\Runner\InlineExecutor">
-            <argument type="service" id="task.handler.factory"/>
+        <service id="task.executor.separate" class="Task\TaskBundle\Executor\SeparateProcessExecutor">
+            <argument type="string">%task.executor.console_path%</argument>
+            <argument type="string">%kernel.environment%%</argument>
         </service>
     </services>
 </container>
