+++ tests/app/config/config.doctrine.yml
@@ -7,13 +7,13 @@ task:
             workload: ['test']
             cron_expression: '* * * * *'
     executor:
-        type: process
-        process:
-            console: '%kernel.root_dir%/console'
+        type: separate
+        separate:
+            console_path: '%kernel.root_dir%/console'
 
 doctrine:
     orm:
-        auto_generate_proxy_classes: %kernel.debug%
+        auto_generate_proxy_classes: '%kernel.debug%'
         auto_mapping: true
     dbal:
         driver: pdo_sqlite
